pyproject.toml
[project]
name = "ambient-listening-backend"
version = "0.1.0"
description = "Backend for the Ambient Listening EMR prototype."
dependencies = [
    "openai-whisper",
    "google-generativeai",
    "python-dotenv",
    "pydantic",
    "fastapi",
    "uvicorn[standard]",
    "python-multipart",
    "pandas>=2.3.1",
    "pdfplumber>=0.11.7",
    "streamlit>=1.47.1",
]
